+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Security
 - Added replace statements to go.mod to remove vulnerable dependency versions from the dependency tree
   [cyberark/sidecar-injector#68](https://github.com/cyberark/sidecar-injector/pull/68)
+  [cyberark/sidecar-injector#69](https://github.com/cyberark/sidecar-injector/pull/69)
 
 ## [0.1.1] - 2020-06-17